refactor to ensure tests are passing
